requirements.txt
numpy==1.26.4
pandas==2.2.2
polars==0.20.29
pip==24.0
pylint==3.2.2
langchain_openai==0.1.8
langchain.chains==0.2.3